% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format}
\alias{format}
\title{Format a dataframe's column names}
\usage{
format(df)
}
\arguments{
\item{df}{A dateframe}
}
\value{
The input dataframe with lower case column names, stripped of punctuation
}
\description{
Format a dataframe's column names
}
\examples{
format(mtcars)
}

